+++ src/BandcampDownloader/Core/LogType.cs
@@ -1,0 +1,11 @@
+﻿namespace BandcampDownloader {
+
+    internal enum LogType {
+        VerboseInfo,
+        Info,
+        IntermediateSuccess,
+        Success,
+        Warning,
+        Error,
+    }
+}