+++ src/BandcampDownloader/Helpers/Exceptions.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace BandcampDownloader {
+
+    internal class CouldNotCheckForUpdatesException: Exception {
+    }
+
+    internal class NoAlbumFoundException: Exception {
+    }
+}