+++ src/BandcampDownloader/Model/JSON/JsonAlbumData.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using Newtonsoft.Json;
+
+namespace BandcampDownloader {
+
+    internal class JsonAlbumData {
+
+        [JsonProperty("title")]
+        public String AlbumTitle { get; set; }
+    }
+}